bot/tests/test_simple.py
#!/usr/bin/env python3
# -*- coding: utf-8 -*-
"""
簡単なテストケース
"""

import pytest
import os
import sys
from unittest.mock import Mock, patch

class TestBasicFunctionality:
    """基本機能のテスト"""

    def test_environment_variables(self):
        """環境変数のテスト"""
        # 必要な環境変数が設定されていることを確認
        required_vars = [
            'DISCORD_BOT_TOKEN_MIYA',
            'DISCORD_BOT_TOKEN_EVE',
            'GCP_PROJECT_ID',
            'GCP_LOCATION'
        ]
        
        with patch.dict(os.environ, {
            'DISCORD_BOT_TOKEN_MIYA': 'test_token_miya',
            'DISCORD_BOT_TOKEN_EVE': 'test_token_eve',
            'GCP_PROJECT_ID': 'test-project',
            'GCP_LOCATION': 'us-central1'
        }):
            for var in required_vars:
                assert os.getenv(var) is not None

    def test_python_version(self):
        """Python バージョンのテスト"""
        # Python 3.9以上であることを確認
        version_info = sys.version_info
        assert version_info.major == 3
        assert version_info.minor >= 9

    def test_import_basic_modules(self):
        """基本モジュールのインポートテスト"""
        # 標準ライブラリのインポート
        import asyncio
        import json
        import datetime
        import os
        
        assert asyncio is not None
        assert json is not None
        assert datetime is not None
        assert os is not None

    def test_import_third_party_modules(self):
        """サードパーティモジュールのインポートテスト"""
        try:
            import discord
            import firebase_admin
            assert discord is not None
            assert firebase_admin is not None
        except ImportError as e:
            pytest.skip(f"サードパーティモジュールが見つかりません: {e}")

    @pytest.mark.asyncio
    async def test_async_functionality(self):
        """非同期機能のテスト"""
        async def sample_async_function():
            return "test_result"
        
        result = await sample_async_function()
        assert result == "test_result"

    def test_mock_functionality(self):
        """モック機能のテスト"""
        mock_object = Mock()
        mock_object.test_method.return_value = "mocked_value"
        
        result = mock_object.test_method()
        assert result == "mocked_value"
        mock_object.test_method.assert_called_once()

    def test_file_operations(self):
        """ファイル操作のテスト"""
        import tempfile
        import json
        
        test_data = {"test": "data", "number": 123}
        
        with tempfile.NamedTemporaryFile(mode='w+', suffix='.json', delete=False) as f:
            json.dump(test_data, f)
            temp_file_path = f.name
        
        try:
            with open(temp_file_path, 'r') as f:
                loaded_data = json.load(f)
            
            assert loaded_data == test_data
        finally:
            os.unlink(temp_file_path)

    def test_character_data_structure(self):
        """キャラクターデータ構造のテスト"""
        character_data = {
            'name': 'みやにゃん',
            'token_env_var': 'DISCORD_BOT_TOKEN_MIYA',
            'emoji': '🐈',
            'personality': 'フレンドリーで好奇心旺盛',
            'speaking_style': 'だにゃ、にゃ〜、だよにゃ',
            'role': '技術解説・コミュニティサポート',
            'color': 0xFF69B4,
            'response_triggers': ['みやにゃん', 'miya', '技術']
        }
        
        assert isinstance(character_data['name'], str)
        assert isinstance(character_data['color'], int)
        assert isinstance(character_data['response_triggers'], list)
        assert len(character_data['response_triggers']) > 0